Overlay-main/api/pyproject.toml
[project]
name = "buildtrace-api"
version = "1.0.0"
description = "BuildTrace API - Central orchestrator for construction drawing comparison"
requires-python = ">=3.12"
dependencies = [
    # Web Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    # Database
    "sqlmodel>=0.0.22",
    "psycopg2-binary>=2.9.10",
    # Storage
    "google-cloud-storage>=2.10.0",
    "boto3>=1.34.0",
    # Message Queue
    "google-cloud-pubsub>=2.23.0",
    # Auth
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    # Configuration
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    # Image Processing (for alignment endpoint)
    "numpy>=1.26.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.3",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.27.0",
    "ruff>=0.8.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP"]
ignore = ["E501"]

[tool.uv]
package = false

